[rename] 클래스,메소드 명수정
