+++ stock/src/main/java/com/sparta/rooibos/stock/domain/repository/StockRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.sparta.rooibos.stock.domain.repository;
 
 
-import com.sparta.rooibos.stock.domain.dto.criteria.StockPageCriteria;
+import com.sparta.rooibos.stock.domain.criteria.StockPageCriteria;
 import com.sparta.rooibos.stock.domain.entity.Pagination;
 import com.sparta.rooibos.stock.domain.entity.Stock;
